Only delete user in ldap is a default conf exists
If there's no default configuration means we are not performing and
synch tasks with ldap by default and we should not worry about removing
users from there.

This bug fix in a way relates with #UL-ISEG-3009 since while trying the
patch for the specific error in that issue we ran into this one.

+++ src/main/java/com/qubit/solution/fenixedu/integration/ldap/servlet/LdapIntegrationInitializer.java
@@ -50,7 +50,9 @@ public class LdapIntegrationInitializer implements ServletContextListener {
 
             @Override
             public void deleting(Person person) {
-                LdapIntegration.deleteUser(person);
+                if (Bennu.getInstance().getDefaultLdapServerIntegrationConfiguration() != null) {
+                    LdapIntegration.deleteUser(person);
+                }
             }
         });
         StudentAccessServices.subscribeSyncPerson(new SyncPersonWithLdap());
